va-langgraph/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "va_langgraph"
version = "0.1.0"
description = "VA assistant — LangGraph (Python)"
readme = "README.md"
requires-python = "==3.12.*"
dependencies = [
    "fastapi>=0.115.0",
    "httpx>=0.28.0",
    "langchain>=0.3.0",
    "langchain-mcp-adapters>=0.1.0",
    "langchain-core>=0.3.0",
    "langchain-google-genai>=2.0.0",
    "langgraph>=0.2.0",
    "langgraph-checkpoint-postgres>=2.0.0",
    "psycopg[binary,pool]>=3.2.0",
    "pydantic>=2.0.0",
    "cachetools>=5.0.0",
    "structlog>=25.0.0",
    "python-dotenv>=1.2.1",
    "uvicorn[standard]>=0.30.0",
]

[tool.hatch.build.targets.wheel]
packages = ["graph", "gateway", "eval"]

[tool.pytest.ini_options]
testpaths = ["tests", "graph"]
pythonpath = ["."]
asyncio_mode = "auto"

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-mock>=3.14.0",
]
notebooks = [
    "marimo>=0.10.0",
    "plotly>=5.0.0",
    "polars>=1.0.0",
]
